Add IMS scenario engine: caller/callee separate-leg recording
2차 확장 1번 — IMS 음성 시나리오 엔진 (전체 83건 통과, ruff clean, IMS 데모 동작).

- scenario/expectations.py: IMS 분기 _build_ims — caller/callee 2레그 파일 기대값
  (I_{callid}_{from}_{to}_ts / 역순), mdn_local(82→0), db_min_rows=1, floor 없음
- scenario/engine.py: _run_ims (INVITE/ACK → recording_start_req → 양레그 RTP →
  recording_stop_req → BYE), _stream_media(emit_floor) 로 MCPTT/IMS 공용화,
  _ims_call_id(실 로그형식 ..._...@ip)
- scenario/loader.py: MediaSpec.duration_sec float 허용(소수초)
- validator/db_check.py: caller/callee 파일명 양쪽 매칭, db_min_rows 사용
- validator/report.py: 골든 파일 조회를 기대 정규식 기반으로(IMS/MCPTT 공용)
- tests/test_scenario_ims.py: 양레그 기대값/엔진 E2E/검증 골든 PASS

run_demo.py --scenario IMS-VOICE-INBOUND: 2레그 500패킷×2, recording_start/stop, 파일 정규식 확인.

+++ backend/sim/scenario/loader.py
@@ -20,7 +20,7 @@ class MediaSpec(BaseModel):
     octet_align: bool = True
     mode_set: list[int] = Field(default_factory=lambda: [8])
     source_audio: Optional[str] = None
-    duration_sec: int = 10
+    duration_sec: float = 10
     pt: int = 98
